Time: 157 ms (28.44%), Space: 63.2 MB (37.91%) - LeetHub

+++ 2616-maximal-score-after-applying-k-operations/2616-maximal-score-after-applying-k-operations.java
@@ -2,15 +2,15 @@ class Solution {
 
   
     public long maxKelements(int[] nums, int k) {
-       PriorityQueue<Double> pq=new PriorityQueue<>(Collections.reverseOrder());
+       PriorityQueue<Integer> pq=new PriorityQueue<>(Collections.reverseOrder());
        for(int i:nums){
-        pq.add((double)i);
+        pq.add(i);
        }
        long ans=0;
        for(int i=0;i<k;i++){
         double popp=pq.remove();
             ans+=popp;
-            pq.add((Math.ceil(popp/3)));
+            pq.add((int)(Math.ceil(popp/3)));
        }
 return ans;
